Finished generalizing a harness and implemented a dolt harness
